python-service/requirements.txt
# ========================================
# Brainstorming Module - Required Packages
# Python 3.10+ required
# ========================================

# ========================================
# Web Framework
# ========================================
fastapi==0.115.0
uvicorn[standard]==0.32.0

# ========================================
# Configuration
# ========================================
python-dotenv==1.0.1
pydantic==2.12.4
pydantic-settings==2.6.1

# ========================================
# OpenAI & LLM
# ========================================
openai==1.57.0

# ========================================
# LangChain & RAG
# ========================================
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.79
langchain-openai==0.2.13
langchain-text-splitters==0.3.11

# ========================================
# Vector Database
# ========================================
chromadb==0.5.23

# ========================================
# Utilities
# ========================================
python-dateutil==2.9.0.post0
httpx==0.27.0
numpy==1.26.4
duckduckgo-search==6.3.0

# ========================================
# Notes
# ========================================
# 브레인스토밍 모듈 전용 (2025-12-09 업데이트)
# - chromadb==0.5.23: 영구 RAG (브레인스토밍 기법)
# - Ephemeral RAG: ChromaDB → JSON 기반으로 전환
# - httpx==0.27.0: 네이버 API 비동기 호출
# - numpy==1.26.4: 코사인 유사도 계산
# - duckduckgo-search==6.3.0: 글로벌 트렌드 검색
#
# 트렌드 검색 모듈 추가:
# - 네이버 뉴스 API
# - 네이버 데이터랩 API  
# - DuckDuckGo 검색 (무료, API 키 불필요)
#
# 제외된 패키지:
# - Database & ORM: 메모리 세션 관리만 사용
# - Authentication: Java Spring Boot에서 처리
# - PDF Processing: 브레인스토밍에서 미사용
# - Testing: 개발용으로 일단 제외
